+++ flink-runtime/src/main/java/org/apache/flink/runtime/io/network/partition/store/tier/dfs/DfsDataManager.java
@@ -135,7 +135,7 @@ public class DfsDataManager implements SingleTierDataGate, DataManagerOperation 
     }
 
     @Override
-    public boolean hasCurrentSegment(int subpartitionId, int segmentIndex) {
+    public boolean hasCurrentSegment(int subpartitionId, long segmentIndex) {
         return segmentIndexTracker.hasCurrentSegment(subpartitionId, segmentIndex);
     }
