+++ flink-runtime/src/main/java/org/apache/flink/runtime/io/network/partition/store/tier/dfs/SubpartitionDfsCacheDataManager.java
@@ -179,11 +179,11 @@ public class SubpartitionDfsCacheDataManager {
         this.outputMetrics = checkNotNull(outputMetrics);
     }
 
-    public void startSegment(int segmentIndex) throws IOException {
+    public void startSegment(long segmentIndex) throws IOException {
         cacheDataSpiller.startSegment(segmentIndex);
     }
 
-    public void finishSegment(int segmentIndex) {
+    public void finishSegment(long segmentIndex) {
         List<ConsumerId> needNotify = new ArrayList<>(consumerMap.size());
         runWithLock(
                 () -> {
